Rename Constants module to constants because it doesn't contain a component
