+++ src/web/components/chart/base/Line.tsx
@@ -27,7 +27,7 @@ import Legend, {
 } from 'web/components/chart/base/Legend';
 import LegendLine from 'web/components/chart/base/LegendLine';
 import Svg from 'web/components/chart/base/Svg';
-import {MENU_PLACEHOLDER_WIDTH} from 'web/components/chart/utils/Constants';
+import {MENU_PLACEHOLDER_WIDTH} from 'web/components/chart/utils/constants';
 import Layout from 'web/components/layout/Layout';
 import PropTypes from 'web/utils/prop-types';
 import Theme from 'web/utils/theme';
