+++ src/web/components/chart/utils/constants.ts
